requirements.txt
asgiref==2.2.0
async-timeout==2.0.1
attrs==17.4.0
autobahn==18.3.1
Automat==0.6.0
bcrypt==3.1.4
certifi==2018.1.18
cffi==1.11.4
channels==2.0.2
chardet==3.0.4
click==6.7
constantly==15.1.0
daphne==2.1.0
diff-match-patch==20121119
Django==1.11.7
django-allauth==0.35.0
django-appconf==1.0.2
django-avatar==4.1.0
django-cors-headers==2.2.0
django-import-export==0.7.0
django-social-share==1.1.2
django-user-agents==0.3.2
django-widget-tweaks==1.4.1
djangorestframework==3.7.7
et-xmlfile==1.0.1
Flask==0.12.2
haversine==0.4.5
hyperlink==18.0.0
idna==2.6
incremental==17.5.0
itsdangerous==0.24
jdcal==1.3
Jinja2==2.10
MarkupSafe==1.0
oauthlib==2.0.6
odfpy==1.3.6
openpyxl==2.5.0
pathlib==1.0.1
Pillow==5.0.0
pycparser==2.18
Pygments==2.2.0
PyJWT==1.6.0
python-openid==2.2.5
pytz==2017.3
PyYAML==3.12
requests==2.18.4
requests-oauthlib==0.8.0
six==1.11.0
social-auth-app-django==2.1.0
social-auth-core==1.7.0
tablib==0.12.1
Twisted==17.9.0
txaio==2.9.0
ua-parser==0.8.0
unicodecsv==0.14.1
urllib3==1.22
user-agents==1.1.0
virtualenv==15.1.0
Werkzeug==0.12.2
xlrd==1.1.0
xlwt==1.3.0
zope.interface==4.4.3
